move option setup functions to module
