+++ main.lua
@@ -164,50 +164,9 @@ local parser_API = require 'modules.apis.parser'
 --------------------------------------------------------------------------------------------------------
 
 local addons = require 'modules.addons'
+local setup = require 'modules.setup'
 
---sets up the compatible extensions list
-local function setup_extensions_list()
-    --setting up subtitle extensions
-    for ext in API.iterate_opt(o.subtitle_extensions:lower()) do
-        sub_extensions[ext] = true
-        extensions[ext] = true
-    end
-
-    --setting up audio extensions
-    for ext in API.iterate_opt(o.audio_extensions:lower()) do
-        audio_extensions[ext] = true
-        extensions[ext] = true
-    end
-
-    --adding file extensions to the set
-    for _, ext in ipairs(compatible_file_extensions) do
-        extensions[ext] = true
-    end
-
-    --adding extra extensions on the whitelist
-    for str in API.iterate_opt(o.extension_whitelist:lower()) do
-        extensions[str] = true
-    end
-
-    --removing extensions that are in the blacklist
-    for str in API.iterate_opt(o.extension_blacklist:lower()) do
-        extensions[str] = nil
-    end
-end
-
---splits the string into a table on the semicolons
-local function setup_root()
-    for str in API.iterate_opt(o.root) do
-        local path = mp.command_native({'expand-path', str})
-        path = API.fix_path(path, true)
-
-        local temp = {name = path, type = 'dir', label = str, ass = API.ass_escape(str, true)}
-
-        root[#root+1] = temp
-    end
-end
-
-setup_root()
+setup.root()
 
 addons.setup_parser(file_parser, "file-browser.lua")
 addons.setup_parser(root_parser, 'file-browser.lua')
@@ -217,7 +176,7 @@ if o.addons then
 end
 
 --these need to be below the addon setup in case any parsers add custom entries
-setup_extensions_list()
+setup.extensions_list()
 keybinds.setup_keybinds()
